изменена модель кор даты, готов экран создания листа, но пока без возможности наполнения продуктами из других листов

+++ GroceryList/Services/DataBase/CoreDataManager.swift
@@ -26,7 +26,7 @@ class CoreDataManager {
         let object = DBGroceryListModel(context: context)
         object.id = list.id
         object.isFavorite = list.isFavorite
-        object.color = list.color
+        object.color = Int64(list.color)
         object.name = list.name
         object.dateOfCreation = list.dateOfCreation
         try? context.save()
@@ -48,7 +48,7 @@ class CoreDataManager {
         if let object = try? context.fetch(fetchRequest).first {
             object.id = list.id
             object.isFavorite = list.isFavorite
-            object.color = list.color
+            object.color = Int64(list.color)
             object.name = list.name
             object.dateOfCreation = list.dateOfCreation
         }
